Introduce GridTools Compiler (gt_toolchain -> gtc) (#38)
