+++ examples/unstructured/simple_reduction_codegen.py
@@ -4,9 +4,9 @@
 from devtools import debug  # noqa: F401
 
 import eve  # noqa: F401
-from gt_toolchain import common
-from gt_toolchain.unstructured import naive
-from gt_toolchain.unstructured.naive_codegen import NaiveCodeGenerator
+from gtc import common
+from gtc.unstructured import naive
+from gtc.unstructured.naive_codegen import NaiveCodeGenerator
 
 
 # ### Reduce edge to node
