+++ setup.cfg
@@ -122,7 +122,7 @@ use_parentheses = true
 lines_after_imports = 2
 default_section = THIRDPARTY
 sections = FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER
-known_first_party = eve,gt_toolchain
+known_first_party = eve,gtc
 known_third_party = black,boltons,devtools,jinja2,mako,packaging,pydantic,pytest,setuptools,sphinx_material,typing_extensions,xxhash
