+++ src/gtc/structured/__init__.py
