+++ src/gtc/structured/common.py
