+++ src/gtc/structured/gt_ir.py
