+++ src/gtc/structured/impl_ir.py
