+++ src/gtc/unstructured/naive_codegen.py
@@ -21,7 +21,7 @@ from typing import ClassVar, Mapping
 from mako import template as mako_tpl
 
 from eve import codegen
-from gt_toolchain import common
+from gtc import common
 
 from .naive import LocationType
