+++ src/gtc/unstructured/sir_passes/infer_local_variable_location_type.py
@@ -16,7 +16,7 @@
 
 import eve  # noqa: F401
 from eve import Node, NodeTranslator, NodeVisitor
-from gt_toolchain import sir
+from gtc import sir
 
 
 class GTException(Exception):
