+++ src/gtc/unstructured/ugpu.py
@@ -20,7 +20,7 @@ from typing import List, Optional, Union
 from pydantic import root_validator
 
 from eve import Node, Str
-from gt_toolchain import common
+from gtc import common
 
 
 class Expr(Node):
