+++ tests/gtc/lit_codegen_suite/lit.cfg
@@ -19,7 +19,7 @@
 import lit.formats
 import os
 
-config.name = "GT_TOOLCHAIN-UNSTRUCTURED"
+config.name = "GTC-UNSTRUCTURED"
 config.test_format = lit.formats.ShTest("0")
 config.suffixes = [".py"]
